+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Improve offenses CLI output format.
+
 ## 0.1.0 - 2021-12-21
 
 - Initial release.